Revert "feat: adds an example with macroquad"
This reverts commit 7eda490b1d515d43246717ec971b036f7afb0436.
