+++ examples/gpui-example/Cargo.toml
@@ -1,8 +1,8 @@
 [package]
 name = "gpui-example"
-version.workspace = true
-edition.workspace = true
+version = "0.1.0"
+edition = "2021"
 
 [dependencies]
-gpui = { git = "https://github.com/zed-industries/zed" }
-lilt = { path = "../.." }
+gpui = { git = "https://github.com/zed-industries/zed", rev = "5b1ea7eda0bba40e222f7d904ec65e5848f45eee" }
+lilt = "0.6.0"
