+++ examples/iced-demo/Cargo.toml
@@ -1,11 +1,11 @@
 [package]
 name = "iced-demo"
-version.workspace = true
-edition.workspace = true
+version = "0.1.0"
+edition = "2021"
 
 [dependencies]
 iced = { git = "https://github.com/iced-rs/iced", rev = "169667ef1b4fa754ed1edb5fa0e845aede2638fb", features = [
     "canvas",
 ] }
-lilt = { path = "../.." }
+lilt = { path = "../../" }
 iced_futures = "0.12.0"
