+++ examples/iced-indicator/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
-name = "iced-indicator"
-version.workspace = true
-edition.workspace = true
+name = "iced-demo"
+version = "0.1.0"
+edition = "2021"
 
 [dependencies]
 iced = { git = "https://github.com/iced-rs/iced", rev = "169667ef1b4fa754ed1edb5fa0e845aede2638fb", features = [
@@ -9,4 +9,4 @@ iced = { git = "https://github.com/iced-rs/iced", rev = "169667ef1b4fa754ed1edb5
     "tokio",
     "svg",
 ] }
-lilt = { path = "../.." }
+lilt = { path = "../../" }
