+++ examples/iced-minimal/Cargo.toml
@@ -1,8 +1,9 @@
 [package]
 name = "iced-minimal"
-version.workspace = true
-edition.workspace = true
+version = "0.1.0"
+edition = "2021"
 
 [dependencies]
 iced = { git = "https://github.com/iced-rs/iced", rev = "169667ef1b4fa754ed1edb5fa0e845aede2638fb" }
-lilt = { path = "../.." }
+lilt = { path = "../../" }
+iced_futures = "0.12.0"
