--- examples/macroquad-example/index.html
@@ -1,28 +1,0 @@
-<html lang="en">
-
-<head>
-    <meta charset="utf-8">
-    <title>Lilt / Macroquad Example</title>
-    <style>
-        html,
-        body,
-        canvas {
-            margin: 0px;
-            padding: 0px;
-            width: 100%;
-            height: 100%;
-            overflow: hidden;
-            position: absolute;
-            background: black;
-            z-index: 0;
-        }
-    </style>
-</head>
-
-<body>
-    <canvas id="glcanvas" tabindex='1'></canvas>
-    <script src="https://not-fl3.github.io/miniquad-samples/mq_js_bundle.js"></script>
-    <script>load("main.wasm");</script>
-</body>
-
-</html>
